% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplot.R
\name{inter_plot}
\alias{inter_plot}
\title{Intersection plot}
\usage{
inter_plot(x, y, nodes = NULL, node_col = "black", node_border = NA,
  edge_col = rgb(0, 0, 0, alpha = 0.5), edge_weight = function(x) if
  (is.null(x)) 1 else x, r = 0.05)
}
\description{
Compares two networks by plotting in a circular layout the respective
networks for only those nodes that are shared across networks.
}
